requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.2
sqlparse==0.4.4
